requirements.txt
# deep learning
torch
lightning
wandb
normflows
git+https://gitlab.cern.ch/mleigh/mltools@projecttemplate

# snakemake script debugging
typer

# config management
hydra-core

# data science & visualization
matplotlib
numpy
scipy
scikit-learn
